+++ test/bacnet-read-test.js
@@ -41,43 +41,43 @@ describe('Read node Testing', function () {
     it('simple read node should be loaded', function (done) {
       helper.load([deviceNode, clientNode, instanceNode, readNode], [
         {
-          "id": "fa0424dc.f9bd",
-          "type": "BACnet-Read",
-          "z": "ad26e8b.6b24498",
-          "name": "bacnetRead",
-          "objectType": "8",
-          "instance": "cf0dca49.2a9ac",
-          "propertyId": "28",
-          "device": "b289851b.dec6f8",
-          "server": "1528f96c.56d047",
-          "multipleRead": false,
-          "wires": [
+          id: 'fa0424dc.f9bd',
+          type: 'BACnet-Read',
+          z: 'ad26e8b.6b24498',
+          name: 'bacnetRead',
+          objectType: '8',
+          instance: 'cf0dca49.2a9ac',
+          propertyId: '28',
+          device: 'b289851b.dec6f8',
+          server: '1528f96c.56d047',
+          multipleRead: false,
+          wires: [
             []
           ]
         },
         {
-          "id": "cf0dca49.2a9ac",
-          "type": "BACnet-Instance",
-          "z": "",
-          "name": "Room Simulator YABE",
-          "instanceAddress": "3342490"
+          id: 'cf0dca49.2a9ac',
+          type: 'BACnet-Instance',
+          z: '',
+          name: 'Room Simulator YABE',
+          instanceAddress: '3342490'
         },
         {
-          "id": "b289851b.dec6f8",
-          "type": "BACnet-Device",
-          "z": "",
-          "name": "Windows VM",
-          "deviceAddress": "192.168.1.94"
+          id: 'b289851b.dec6f8',
+          type: 'BACnet-Device',
+          z: '',
+          name: 'Windows VM',
+          deviceAddress: '192.168.1.94'
         },
         {
-          "id": "1528f96c.56d047",
-          "type": "BACnet-Client",
-          "z": "",
-          "name": "",
-          "adpuTimeout": "",
-          "port": "",
-          "interface": "",
-          "broadcastAddress": ""
+          id: '1528f96c.56d047',
+          type: 'BACnet-Client',
+          z: '',
+          name: '',
+          adpuTimeout: '',
+          port: '',
+          interface: '',
+          broadcastAddress: ''
         }
       ], function () {
         var bacnetRead = helper.getNode('fa0424dc.f9bd')